Fix no subscriber error message initial pop up
